Clarified error to be more descriptive of the problem.
The error condition being signaled here is not that the world (WRLD
record) isn't found in the ESM file, but that there are no LAND records
associated with that world.  The absence of the world itself is checked
for and reported elsewhere, but the same error text was being used in
both places.